switch to map_str to avoid wrong domain matching
we have many entries in the backend.map file, and some are subdomains of
other entries, which should map to other backends. Using the map_dom
function means that sometimes a domain is matched incorrectly. For
example, a request to `foo.sandbox.opencraft.hosting` could match a
mapping for `sandbox.opencraft.hosting` in `backend.map`, even if there
was a `foo.sandbox.opencraft.hosting` entry in the mapping.