Fix a bug ignoring the 1st certificate when rendering certificates
The code to ignore the first line in the output of the consul command
to get the keys was added because there was an empty 'ocim/certs/' key
in the staging environment where this was tested. But the key wasn't
present in the production environment, thereby causing the first
certificate to be ignored. This fix will work in the stage environment
after removing the empty key there.