Remove listen on port 80 for mailman3 nginx proxy
Only the common-server should listen on port 80,
because it uses that to intercept requests to the well-known acme-challenge location
for certbot renewals.
This server block was blocking certbot renewals by sending the acme-challenge
requests to mailman, instead of letting the common-server server block handle it.

By removing this server block, we fix certbot renewals on the mail server instance.