Run the sanity checker every 10 minutes
Previously, it was _supposed_ to run every 10 minutes,
but it skipped '0', so there was a 20min gap every hour,
which can cause false positives for snitches.

Also don't use the `*/10` syntax,
because we want to offset it by one minute,
to avoid issues where everything happens on a minute divisible by 10.