Update filebeat configuration for version 7
Rename `prospectors` to `inputs`.
`prospectors` was deprecated in version 6.3, and removed in a later version,
so we can rename without breaking any hosts now (all seem to be 6.8 or 7.*).

Remove `registry_file`.
This was a backwards incompatible change somewhere between version 6.8 and 7.15.
We aren't overriding the default anywhere, so we can simply remove the option altogether.